feat: add cjsToEsm.cjs solution
