+++ src/modules/esm.mjs
@@ -1,0 +1,46 @@
+import { dirname } from 'path';
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+import { promises as fs } from 'fs';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+const a = await fs.readFile(path.join(__dirname, 'files', 'a.json'), 'utf-8');
+const b = await fs.readFile(path.join(__dirname, 'files', 'b.json'), 'utf-8');
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = a;
+} else {
+    unknownObject = b;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.url}`);
+console.log(`Path to current directory is ${path.dirname(import.meta.url)}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export default {
+    unknownObject,
+    myServer,
+};
+
